requirements.txt
spotipy
libresonic
python-dotenv
